+++ shuup/admin/modules/products/forms/package_forms.py
@@ -10,18 +10,32 @@ from __future__ import unicode_literals
 import six
 from django import forms
 from django.contrib import messages
+from django.db.models import ObjectDoesNotExist
 from django.db.transaction import atomic
 from django.utils.translation import ugettext_lazy as _
 
 from shuup.admin.forms.widgets import PackageProductChoiceWidget
 from shuup.admin.modules.products.utils import clear_existing_package
 from shuup.core.excs import ImpossibleProductModeException, Problem
-from shuup.core.models import Product
+from shuup.core.models import Product, Shop
 
 from .parent_forms import ProductChildBaseFormSet
 
 
 class PackageChildForm(forms.Form):
+    def __init__(self, **kwargs):
+        initial = kwargs.get("initial", {})
+        self.product = initial.get("child")
+        if self.product:
+            self.shop_products = []
+            for shop in Shop.objects.all():
+                try:
+                    shop_product = self.product.get_shop_instance(shop)
+                    self.shop_products.append(shop_product)
+                except ObjectDoesNotExist:
+                    continue
+        super(PackageChildForm, self).__init__(**kwargs)
+
     child = forms.ModelChoiceField(
         queryset=Product.objects.all(),
         widget=PackageProductChoiceWidget(),
@@ -35,9 +49,45 @@ class PackageChildForm(forms.Form):
         required=True,
     )
 
+    def get_stock_statuses(self):
+        stocks = {}
+        if not self.product:
+            return stocks
+        sales_unit = self.product.sales_unit
+        sales_decimals = sales_unit.decimals if sales_unit else 0
+        sales_unit_short_name = sales_unit.short_name if sales_unit else ""
+        for shop_product in self.shop_products:
+            for supplier in shop_product.suppliers.all():
+                if supplier in stocks.keys():
+                    continue
+                stock_status = supplier.get_stock_status(product_id=self.product.id)
+                stocks[supplier] = (supplier, stock_status, sales_decimals, sales_unit_short_name)
+        return stocks
+
+    def get_orderability_errors(self):
+        orderability_errors = []
+        if not self.product:
+            return orderability_errors
+        for shop_product in self.shop_products:
+            orderability_errors.extend(
+                ["%s: %s" % (shop_product.shop.name, msg.message)
+                 for msg in shop_product.get_orderability_errors(
+                    supplier=None,
+                    quantity=shop_product.minimum_purchase_quantity,
+                    customer=None)]
+            )
+            for supplier in shop_product.suppliers.all():
+                orderability_errors.extend(
+                    ["%s: %s" % (supplier.name, msg.message)
+                     for msg in supplier.get_orderability_errors(
+                        shop_product=shop_product,
+                        quantity=shop_product.minimum_purchase_quantity,
+                        customer=None)])
+        return orderability_errors
+
 
 class PackageChildFormSet(ProductChildBaseFormSet):
-    deletion_label = _("Remove")
+    deletion_label = ""
 
     def __init__(self, **kwargs):
         self.parent_product = kwargs.pop("parent_product")